backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "colorama>=0.4.6",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "orjson>=3.11.5",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "argon2-cffi>=25.1.0",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "sqlacodegen>=3.2.0",
]
